+++ src/Serializer/tests/Serializer/ProtoSerializerTest.php
@@ -4,7 +4,6 @@ declare(strict_types=1);
 
 namespace Spiral\Tests\Serializer\Serializer;
 
-use Google\Protobuf\Internal\Message;
 use PHPUnit\Framework\Attributes\DataProvider;
 use PHPUnit\Framework\TestCase;
 use Spiral\Serializer\Exception\InvalidArgumentException;
@@ -20,19 +19,13 @@ final class ProtoSerializerTest extends TestCase
         $this->serializer = new ProtoSerializer();
     }
 
-    public function testSerialize(): void
-    {
-        $payload = $this->createMock(Message::class);
-        $payload->expects($this->once())->method('serializeToString')->willReturn('serialized');
-
-        $this->assertSame('serialized', $this->serializer->serialize($payload));
-    }
-
     public function testSerializeMessage(): void
     {
+        $message = new PingRequest(['url' => 'foo']);
+
         $this->assertSame(
-            (new PingRequest(['url' => 'foo']))->serializeToString(),
-            $this->serializer->serialize(new PingRequest(['url' => 'foo']))
+            $message->serializeToString(),
+            $this->serializer->serialize($message)
         );
     }
 
@@ -44,6 +37,7 @@ final class ProtoSerializerTest extends TestCase
             $message,
             $this->serializer->unserialize($message->serializeToString(), PingRequest::class)
         );
+
         $this->assertEquals(
             $message,
             $this->serializer->unserialize($message->serializeToString(), new PingRequest())